pyproject.toml
[build-system]
requires = ["setuptools<69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "instructionkit"
version = "0.1.0"
description = "CLI tool for managing AI coding tool instructions"
readme = "README.md"
authors = [{ name = "Troy Larson", email = "troy@calvinware.com" }]
requires-python = ">=3.10"
license = { text = "MIT License" }
keywords = ["cli", "ai", "instructions", "cursor", "copilot", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "textual>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "invoke>=2.0.0",
    "build>=1.2.1",
    "twine>=5.0.0",
]

[project.scripts]
inskit = "instructionkit.cli.main:app"

[project.urls]
Homepage = "https://github.com/instructionkit/instructionkit"
Documentation = "https://github.com/instructionkit/instructionkit#readme"
Repository = "https://github.com/instructionkit/instructionkit"
Issues = "https://github.com/instructionkit/instructionkit/issues"

[tool.setuptools]
packages = ["instructionkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=instructionkit --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
